pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ims"
version = "0.1.0"
description = "FastAPI microservice for a thrift shop inventory, sales, payments, and analytics."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.0",
    "SQLAlchemy==2.0.29",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

